requirements.txt
Flask==2.1.1
pymongo==4.1.1
requests==2.27.1
pyjwt==1.4.2